+++ CHANGELOG.md
@@ -17,6 +17,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - ids parameter to transaction.query
 - ids parameter to transfer.query
 
+### Fixed
+- UtilityPayment test case
+
 ## [0.6.0] - 2020-08-20
 ### Added
 - transfer->scheduled parameter to allow Transfer scheduling